fix: rename to qwen2.py
