+++ tpu_commons/models/jax/model_loader.py
@@ -20,7 +20,7 @@ def _get_model_architecture(config: PretrainedConfig) -> nn.Module:
         from tpu_commons.models.jax.llama_nn import LlamaForCausalLM
     elif impl == "flax_nnx":
         from tpu_commons.models.jax.llama import LlamaForCausalLM
-        from tpu_commons.models.jax.qwen import Qwen2ForCausalLM
+        from tpu_commons.models.jax.qwen2 import Qwen2ForCausalLM
     else:
         raise NotImplementedError("Unsupported MODEL_IMPL_TYPE")
